// FrontEnd_Locator/src/app/models/ruta-geo.model.ts
export interface RutaGeoModel {
  id: number;
  tienda: string;
  horaSalidaCentroDistribucion: string;
  etaactual: string;
  etainicio: string;
  numeroGuia: string;
  estado: string;
  usuarioCreacion: string;
  latitudOrigen: number;
  longitudOrigen: number;
  latitudDestino: number;
  longitudDestino: number;
}